aws/lambdas/meme-renderer/requirements.txt
Pillow>=10.2.0
boto3>=1.34.0
supabase>=2.4.0
